LayoutTests/imported/w3c/web-platform-tests/trusted-types/TrustedTypePolicyFactory-createPolicy-cspTests-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'trusted-types'.


PASS Allowed-name policy creation works.
PASS Another allowed-name policy creation works.
FAIL Non-allowed name policy creation throws. assert_throws_js: function "_ => {
     window.trustedTypes.createPolicy('SomeOtherName', { createURL: s => s } );
    }" did not throw
